combine config utils, allow global access
