+++ test/index.js
@@ -7,7 +7,7 @@ process.chdir('./fixtures')
 
 require('./index.test').default(test, assert)
 require('./load.test').default(test, assert)
-require('./createConfigFromCLI.test').default(test, assert)
+require('./config.test').default(test, assert)
 require('./createHeadTags.test').default(test, assert)
 require('./encodeFilename.test').default(test, assert)
 require('./createEntries.test').default(test, assert)
